requirements.txt
fastapi
uvicorn
pydantic
pillow
torchvision
pytest